Implement PositionFragment to an almost complete level Add localization parser, update Mod definition to contain mod localization Make small changes to parser types to return Double rather than Float or Number

+++ src/main/kotlin/com/zelkatani/model/map/Default.kt
@@ -29,7 +29,7 @@ data class Default(
     val borderHeights: List<Int>,
     val terrainSheetHeights: Int,
     val tree: Int,
-    val borderCutoff: Float
+    val borderCutoff: Double
 ) {
     companion object : ModelBuilder<Default> {
         override fun from(file: File): Default {
